requirements.txt
pyyaml
requests
pandas
web3
pymongo
dnspython